src_test/tool_test/selenium/25-31/requirements.txt
selenium==4.21.0
webdriver-manager==4.0.1
pytest==8.2.0
pytest-html==4.1.1
mysql-connector-python==9.0.0
openpyxl==3.1.2
Pillow==10.3.0
requests==2.31.0
